move prm to proofs
